RPI_GUI/RPI_GUI_Transmitter/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rpi-gui-app"
version = "0.1.0"
description = "PySide6 GUI for controlling camera and PWM on Raspberry Pi"
authors = [
    { name="Stephano"}
]
requires-python = ">=3.8"

dependencies = [
    "PySide6>=6.5",
    "pyserial"
]

[tool.setuptools]
packages = ["controllers", "windows", "utils"]
